flux-lora-trainer/requirements.txt
# Flux LoRA Training Requirements

# Core dependencies
torch>=2.1.0
torchvision>=0.16.0
diffusers>=0.27.0
transformers>=4.38.0
accelerate>=0.27.0
peft>=0.9.0
safetensors>=0.4.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Utilities
huggingface-hub>=0.20.0
wandb>=0.16.0  # Optional: for experiment tracking
omegaconf>=2.3.0
tqdm>=4.66.0
numpy>=1.24.0

# Memory optimization
bitsandbytes>=0.41.0  # For 8-bit optimizers
xformers>=0.0.22  # Optional: for memory-efficient attention

# Captioning (optional)
sentencepiece>=0.1.99

# CLIP for image understanding
ftfy>=6.1.0
regex>=2023.0.0

# Configuration
pyyaml>=6.0.0
